# source/mobile-beta-foundation/services/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Athletica AI API
  version: 0.5.0
  description: |
    Phase 1 surface of the Athletica AI backend: liveness, authentication with
    logout and token revocation, the idempotent workout-set write, the workout
    lifecycle, the workout history behind the "Итоги" screen and the aggregates
    behind the "Прогресс" screen.

    Three rules hold across the whole contract:

    * the caller's user ID always comes from the bearer access token, never from
      a request body, a query parameter or a cursor;
    * a workout that belongs to another user answers exactly like a workout that
      does not exist (`404`), so existence never leaks;
    * a state change the lifecycle forbids answers `409 invalid_transition` —
      never `500`, and never a silent no-op.

    ### Changes in 0.5.0 (all additive, 0.4.0 clients keep working)

    * `PATCH /workouts/{workoutId}/sets/{setId}` corrects a logged set;
    * `DELETE /workouts/{workoutId}/sets/{setId}` removes one, softly;
    * `PATCH /workouts/{workoutId}` names a workout after it was started —
      the session begun offline with no title is what it exists for;
    * `WorkoutSet` gained the optional `updatedAt` and `deletedAt` fields;
    * error codes `set_deleted` and `workout_not_editable` were added;
    * metrics are exported in Prometheus text form on a **separate** listener
      (`ATHLETICA_METRICS_ADDR`, loopback by default, `ATHLETICA_METRICS_TOKEN`
...

...

...

...

...

...

...
    * `GET /progress`;
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /workouts/{workoutId}/sets:
    parameters:
      - name: workoutId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [workouts]
      summary: Log one strength-training set idempotently
      description: |
        The set is stored under the user ID carried by the access token; a
        `userId` field in the body is ignored.

        Idempotency rests on a unique index on `(user_id, client_mutation_id)`
        in PostgreSQL, so a retried outbox entry can never create a second row,
        even when two requests race. Replaying a mutation ID returns `409` with
        the originally stored set — including when the replayed payload differs.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkoutSetInput' }
      responses:
        '201':
          description: Set logged
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkoutSet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/WorkoutNotFound' }
        '409':
          description: |
            Duplicate client mutation. The body carries the set that was stored
            when this clientMutationId was first accepted.
          content:
            application/json:
              schema:
                type: object
                required: [error, set]
                properties:
                  error: { $ref: '#/components/schemas/ErrorPayload' }
                  set: { $ref: '#/components/schemas/WorkoutSet' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    get:
      tags: [workouts]
      summary: List the sets of one of the caller's workouts
      responses:
        '200':
          description: Sets of the workout, ordered by set number
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/WorkoutSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/WorkoutNotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived HS256 access token issued by /auth/login.
  responses:
    BadRequest:
      description: The body is not a single well-formed JSON object
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, malformed or expired access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: The payload violates the shared domain rules
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    WorkoutNotFound:
      description: |
        The workout does not exist, or belongs to another user. The two cases
        are indistinguishable by design.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    WorkoutSetInput:
      type: object
      required: [exerciseId, setNumber, weightKg, repetitions, rir, clientMutationId]
      description: |
        Bounds mirror validateSet() in packages/domain/src/workout.ts. Any
        additional property — `userId` in particular — is ignored by the server.
      properties:
        exerciseId: { type: string, minLength: 1, maxLength: 128 }
        setNumber: { type: integer, minimum: 1 }
        weightKg: { type: number, minimum: 0, maximum: 1000 }
        repetitions: { type: integer, minimum: 1, maximum: 100 }
        rir: { type: integer, minimum: 0, maximum: 10 }
        clientMutationId:
          type: string
          minLength: 1
          maxLength: 128
          description: Stable per-mutation ID from the client outbox.
...
    WorkoutSet:
      type: object
      required: [id, workoutId, exerciseId, setNumber, weightKg, repetitions, rir, clientMutationId, createdAt, updatedAt, deletedAt]
      properties:
        id: { type: string, format: uuid }
        workoutId: { type: string, format: uuid }
        exerciseId: { type: string }
        setNumber:
          type: integer
          description: |
            The number the client gave the set, never reassigned by the server.
            Gaps are normal after a deletion.
        weightKg: { type: number }
        repetitions: { type: integer }
        rir: { type: integer }
        clientMutationId:
          type: string
          description: |
            The ID of the mutation that *created* the set. A correction does not
            replace it.
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: |
            When the set was last corrected; equal to `createdAt` for a set that
            was never edited. Added in 0.5.0.
        deletedAt:
          type: [string, 'null']
          format: date-time
          description: |
            When the athlete removed the set; `null` for every set a client can
            still see. It is non-null only in the answer to a deletion and in
            the `409` a replayed creation of a removed set gets. Added in 0.5.0.
...
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - unauthorized
            - invalid_credentials
            - email_taken
            - not_found
            - duplicate_client_mutation
            - invalid_transition
            - invalid_cursor
            - workout_id_taken
            - set_deleted
            - workout_not_editable
            - rate_limited
            - internal_error
        message: { type: string }
        details:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorPayload' }